requirements.txt
python-telegram-bot==20.7
pytz
aiogram==2.25.2
apscheduler 